manager/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pct-manager"
version = "0.1.0"
description = "Postgres Control Tower — Manager (FastAPI hub for fleets of pgBackRest-managed PostgreSQL clusters)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pydantic[email]>=2.6",
  "pydantic-settings>=2.2",
  "passlib[bcrypt]>=1.7",
  "pyjwt>=2.8",
  "python-multipart>=0.0.9",
  "apscheduler>=3.10",
  "httpx>=0.27",
]

[project.scripts]
pct-manager = "pct_manager.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["pct_manager"]
